Add files via upload
This is a simple code using HTML and CSS to design the Frontend part of a website. To make the website easy to navigate, a box (Start) to start providing instructions to the robot was added. As soon as this button pressed, the user would have full control over the robot.