Add files via upload
This is the code for the sub-page created in the website to allow full control over the robot. The buttons were organized depending on their direction as simple icons with written part were included. The purpose of this was to make it easier for people at different ages to navigate in the website.